changed main benchmark to only write output into 1 file and added timestamp to file for later sorting

+++ libraries/main_benchmark.py
@@ -7,7 +7,7 @@ sys.path.insert(1, os.path.join(sys.path[0], '..'))
 from benchmark_framework import benchmarker, metrics, dataset_base
 from utils.lcer import get_logger, get_output_result_data, get_base_path
 from utils.filer import write_data, append_data
-from datetime import date
+import time
 from data_processing.preprocessing.preprocessing_100k import load_ml_100k
 from libraries.name_lib_mapping import NAME_LIB_MAP
 import pandas as pd
@@ -19,6 +19,8 @@ logger = get_logger("BenchmarkExe")
 
 # Read Input (the lib name to run)
 lib_name = str(sys.argv[1])
+fresh_start = False
+
 # Load algos from lib name
 lib_algos = NAME_LIB_MAP[lib_name]()
 
@@ -26,12 +28,15 @@ lib_algos = NAME_LIB_MAP[lib_name]()
 dataset_load_functions = [load_ml_100k]
 nr_datasets = len(dataset_load_functions)
 
-#  File management
-output_filepath = os.path.join(get_base_path(), get_output_result_data(), "{}_{}_overall_benchmark_results.csv".format(
-    date.today(), lib_name))
-if os.path.isfile(output_filepath):
+# ------------- File management
+output_filepath = os.path.join(get_base_path(), get_output_result_data(), "overall_benchmark_results.csv")
+
+if fresh_start and os.path.isfile(output_filepath):
     os.remove(output_filepath)
-write_data(pd.DataFrame([], columns=["Dataset", "Model", "LibraryCategory", "RSME", "TimeInSeconds"]), output_filepath)
+
+if not os.path.isfile(output_filepath) or fresh_start:
+    write_data(pd.DataFrame([], columns=["timestamp", "Dataset", "Model", "LibraryCategory", "RSME", "TimeInSeconds"]),
+               output_filepath)
 
 # ------------- Loop over all datasets
 logger.info("######## Loop over all Datasets and do benchmarks for library {} ########".format(lib_name))
@@ -48,15 +53,16 @@ for idx, dataset_function in enumerate(dataset_load_functions, 1):
 
     # Execute benchmarks for this dataset
     for benchmark in benchmarks:
-        tmp_result_data = [(dataset.name, benchmark.model.name, benchmark.model.library_category, *benchmark.run())]
+        tmp_result_data = [(time.time(),dataset.name, benchmark.model.name, benchmark.model.library_category,
+                            *benchmark.run())]
         logger.info("###### Intermediate Result Output and Saving Data ######")
-        _, model_name, _, metric_val, execution_time = tmp_result_data[0]
+        _, _, model_name, _, metric_val, execution_time = tmp_result_data[0]
         logger.info("{}: RSME of {} | Time take {}".format(model_name, metric_val, execution_time))
 
         # Build tmp df to output data
         append_data(
-            pd.DataFrame(tmp_result_data, columns=["Dataset", "Model", "LibraryCategory", "RSME", "TimeInSeconds"]),
-            output_filepath)
+            pd.DataFrame(tmp_result_data, columns=["datetime", "Dataset", "Model", "LibraryCategory", "RSME",
+                                                   "TimeInSeconds"]), output_filepath)
 
     logger.info("###### Finished processing Dataset {} ######".format(dataset.name))
